refactor!: move themes to Themes
misc:
- set MainView.axaml encoding to UTF-8-BOM

BREAKING CHANGES:
- GroupBox themes' paths have changed
