+++ GroupBox.Avalonia/Themes/ModernTheme.axaml
@@ -1,4 +1,4 @@
-<Styles x:Class="GroupBox.Avalonia.ControlThemes.ModernTheme"
+﻿<Styles x:Class="GroupBox.Avalonia.Themes.ModernTheme"
         xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:ctrl="using:GroupBox.Avalonia.Controls">
